+++ lib/BitcoinCash.ts
@@ -9,31 +9,6 @@ const wif = require("wif")
 
 const Buffer = require("safe-buffer").Buffer
 
-export interface BitcoinCash {
-  _address: any
-  toSatoshi(coins: number): number
-  toBitcoinCash(satoshis: number): number
-  signMessageWithPrivKey(privateKeyWIF: string, message: string): string
-  verifyMessage(address: string, signature: string, message: string): boolean
-  encodeBase58Check(hex: string): string
-  decodeBase58Check(address: string): string
-  encodeBIP21(
-    address: string,
-    options: EncodeBIP21Options,
-    regtest?: boolean
-  ): string
-  decodeBIP21(url: string): BIP21Object
-  getByteCount(
-    inputs: ByteCountInput | object,
-    outputs: ByteCountOutput
-  ): number
-  encryptBIP38(privKeyWIF: string, passphrase: string): string
-  decryptBIP38(
-    encryptedKey: string,
-    passphrase: string,
-    network: string
-  ): string
-}
 
 export interface EncodeBIP21Options {
   amount?: number
@@ -55,7 +30,7 @@ export interface ByteCountOutput {
   P2SH?: number
 }
 
-export class BitcoinCash implements BitcoinCash {
+export class BitcoinCash {
   _address: any
   constructor(address: any) {
     this._address = address
