+++ lib/ECPair.ts
@@ -1,36 +1,15 @@
+import * as bcl from "bitcoincashjs-lib";
+
 const Bitcoin = require("bitcoincashjs-lib")
 const coininfo = require("coininfo")
-import { Buffer } from "buffer"
-export type ECSignature = any
 
-export interface ECPair {
-  _address: any
-  fromWIF(privateKeyWIF: string): ECPair
-  toWIF(): string
-  toWIF(ecpair: ECPair): string
-  sign(sigHash: number): ECSignature
-  sign(ecpair: ECPair, sigHash: number): ECSignature
-  sign(buffer: Buffer): Boolean | ECSignature
-  sign(ecpair: ECPair, buffer: Buffer): Boolean | ECSignature
-  verify(buffer: Buffer, signature: ECSignature): boolean
-  verify(ecpair: ECPair, buffer: Buffer, signature: ECSignature): boolean
-  fromPublicKey(pubkeyBuffer: Buffer): ECPair
-  toPublicKey(): Buffer
-  toPublicKey(ecpair: ECPair): Buffer
-  getPublicKeyBuffer(): Buffer
-  toLegacyAddress(): string
-  toLegacyAddress(ecpair: ECPair): string
-  toCashAddress(): string
-  toCashAddress(ecpair: ECPair, regtest?: boolean): string
-  getAddress(): string
-}
-export class ECPair implements ECPair {
+export class ECPair {
   static _address: any
   static setAddress(address: any): void {
     ECPair._address = address
   }
 
-  static fromWIF(privateKeyWIF: string): any {
+  static fromWIF(privateKeyWIF: string): bcl.ECPair {
     let network: string = "mainnet"
     if (privateKeyWIF[0] === "L" || privateKeyWIF[0] === "K")
       network = "mainnet"
@@ -45,31 +24,31 @@ export class ECPair implements ECPair {
     return Bitcoin.ECPair.fromWIF(privateKeyWIF, bitcoincashBitcoinJSLib)
   }
 
-  static toWIF(ecpair: ECPair): string {
+  static toWIF(ecpair: bcl.ECPair): string {
     return ecpair.toWIF()
   }
 
-  static sign(ecpair: ECPair, buffer: Buffer): ECSignature {
+  static sign(ecpair: bcl.ECPair, buffer: Buffer): bcl.ECSignature {
     return ecpair.sign(buffer)
   }
 
-  static verify(ecpair: ECPair, buffer: Buffer, signature: ECSignature): any {
+  static verify(ecpair: bcl.ECPair, buffer: Buffer, signature: bcl.ECSignature): boolean {
     return ecpair.verify(buffer, signature)
   }
 
-  static fromPublicKey(pubkeyBuffer: Buffer): any {
+  static fromPublicKey(pubkeyBuffer: Buffer): bcl.ECPair {
     return Bitcoin.ECPair.fromPublicKeyBuffer(pubkeyBuffer)
   }
 
-  static toPublicKey(ecpair: ECPair): Buffer {
+  static toPublicKey(ecpair: bcl.ECPair): Buffer {
     return ecpair.getPublicKeyBuffer()
   }
 
-  static toLegacyAddress(ecpair: ECPair): string {
+  static toLegacyAddress(ecpair: bcl.ECPair): string {
     return ecpair.getAddress()
   }
 
-  static toCashAddress(ecpair: ECPair, regtest: boolean = false): string {
+  static toCashAddress(ecpair: bcl.ECPair, regtest: boolean = false): string {
     return ECPair._address.toCashAddress(ecpair.getAddress(), true, regtest)
   }
 }
