+++ lib/Generating.ts
@@ -1,15 +1,6 @@
 import axios from "axios"
 
-export interface Generating {
-  restURL: string
-  generateToAddress(
-    blocks: number,
-    address: string,
-    maxtries?: number
-  ): Promise<any[]>
-}
-
-export class Generating implements Generating {
+export class Generating {
   restURL: string
   constructor(restURL: string) {
     this.restURL = restURL
@@ -19,7 +10,7 @@ export class Generating implements Generating {
     blocks: number,
     address: string,
     maxtries: number = 1000000
-  ): Promise<any[]> {
+  ): Promise<string[]> {
     try {
       const response: any = await axios.post(
         `${
