+++ lib/HDNode.ts
@@ -1,8 +1,7 @@
 const Bitcoin = require("bitcoincashjs-lib")
 const coininfo = require("coininfo")
 const bip32utils = require("bip32-utils")
-import { ECPair, ECSignature } from "./ECPair"
-import * as bitcoincashjs from "bitcoincashjs-lib"
+import * as bcl from "bitcoincashjs-lib"
 
 export class HDNode {
   _address: any
@@ -10,7 +9,7 @@ export class HDNode {
     this._address = address
   }
 
-  fromSeed(rootSeedBuffer: any, network: string = "mainnet"): HDNode {
+  fromSeed(rootSeedBuffer: any, network: string = "mainnet"): bcl.HDNode {
     let bitcoincash: any
     if (network === "bitcoincash" || network === "mainnet")
       bitcoincash = coininfo.bitcoincash.main
@@ -23,35 +22,35 @@ export class HDNode {
     )
   }
 
-  toLegacyAddress(hdNode: bitcoincashjs.HDNode): string {
+  toLegacyAddress(hdNode: bcl.HDNode): string {
     return hdNode.getAddress()
   }
 
-  toCashAddress(hdNode: bitcoincashjs.HDNode, regtest: boolean = false): string {
+  toCashAddress(hdNode: bcl.HDNode, regtest: boolean = false): string {
     return this._address.toCashAddress(hdNode.getAddress(), true, regtest)
   }
 
-  toWIF(hdNode: bitcoincashjs.HDNode): string {
+  toWIF(hdNode: bcl.HDNode): string {
     return hdNode.keyPair.toWIF()
   }
 
-  toXPub(hdNode: bitcoincashjs.HDNode): string {
+  toXPub(hdNode: bcl.HDNode): string {
     return hdNode.neutered().toBase58()
   }
 
-  toXPriv(hdNode: bitcoincashjs.HDNode): string {
+  toXPriv(hdNode: bcl.HDNode): string {
     return hdNode.toBase58()
   }
 
-  toKeyPair(hdNode: bitcoincashjs.HDNode): ECPair {
+  toKeyPair(hdNode: bcl.HDNode): bcl.ECPair {
     return hdNode.keyPair
   }
 
-  toPublicKey(hdNode: bitcoincashjs.HDNode): Buffer {
+  toPublicKey(hdNode: bcl.HDNode): Buffer {
     return hdNode.getPublicKeyBuffer()
   }
 
-  fromXPriv(xpriv: string): bitcoincashjs.HDNode {
+  fromXPriv(xpriv: string): bcl.HDNode {
     let bitcoincash: any
     if (xpriv[0] === "x") bitcoincash = coininfo.bitcoincash.main
     else if (xpriv[0] === "t") bitcoincash = coininfo.bitcoincash.test
@@ -60,7 +59,7 @@ export class HDNode {
     return Bitcoin.HDNode.fromBase58(xpriv, bitcoincashBitcoinJSLib)
   }
 
-  fromXPub(xpub: string): bitcoincashjs.HDNode {
+  fromXPub(xpub: string): bcl.HDNode {
     let bitcoincash: any
     if (xpub[0] === "x") bitcoincash = coininfo.bitcoincash.main
     else if (xpub[0] === "t") bitcoincash = coininfo.bitcoincash.test
@@ -69,35 +68,35 @@ export class HDNode {
     return Bitcoin.HDNode.fromBase58(xpub, bitcoincashBitcoinJSLib)
   }
 
-  derivePath(hdnode: bitcoincashjs.HDNode, path: string): bitcoincashjs.HDNode {
+  derivePath(hdnode: bcl.HDNode, path: string): bcl.HDNode {
     return hdnode.derivePath(path)
   }
 
-  derive(hdnode: bitcoincashjs.HDNode, path: string): bitcoincashjs.HDNode {
+  derive(hdnode: bcl.HDNode, path: string): bcl.HDNode {
     return hdnode.derive(path)
   }
 
-  deriveHardened(hdnode: bitcoincashjs.HDNode, path: string): bitcoincashjs.HDNode {
+  deriveHardened(hdnode: bcl.HDNode, path: string): bcl.HDNode {
     return hdnode.deriveHardened(path)
   }
 
-  sign(hdnode: bitcoincashjs.HDNode, buffer: Buffer): ECSignature {
+  sign(hdnode: bcl.HDNode, buffer: Buffer): bcl.ECSignature {
     return hdnode.sign(buffer)
   }
 
-  verify(hdnode: bitcoincashjs.HDNode, buffer: Buffer, signature: ECSignature): boolean {
+  verify(hdnode: bcl.HDNode, buffer: Buffer, signature: bcl.ECSignature): boolean {
     return hdnode.verify(buffer, signature)
   }
 
-  isPublic(hdnode: bitcoincashjs.HDNode): boolean {
+  isPublic(hdnode: bcl.HDNode): boolean {
     return JSON.parse(hdnode.isNeutered())
   }
 
-  isPrivate(hdnode: bitcoincashjs.HDNode): boolean {
+  isPrivate(hdnode: bcl.HDNode): boolean {
     return !hdnode.isNeutered()
   }
 
-  toIdentifier(hdnode: bitcoincashjs.HDNode): string {
+  toIdentifier(hdnode: bcl.HDNode): string {
     return hdnode.getIdentifier()
   }
 
@@ -105,14 +104,14 @@ export class HDNode {
     return Bitcoin.HDNode.fromBase58(base58, network)
   }
 
-  createAccount(hdNodes: bitcoincashjs.HDNode[]): object {
+  createAccount(hdNodes: bcl.HDNode[]): object {
     const arr = hdNodes.map(
       (item: any, index: number) => new bip32utils.Chain(item.neutered())
     )
     return new bip32utils.Account(arr)
   }
 
-  createChain(hdNode: bitcoincashjs.HDNode): object {
+  createChain(hdNode: bcl.HDNode): object {
     return new bip32utils.Chain(hdNode)
   }
 }
