+++ lib/Mining.ts
@@ -1,14 +1,7 @@
 import axios from "axios"
 
-export interface Mining {
-  restURL: string
-  getBlockTemplate(template_request: any): Promise<any>
-  getMiningInfo(): Promise<any>
-  getNetworkHashps(nblocks?: number, height?: number): Promise<number>
-  submitBlock(hex: string, parameters: any): Promise<any>
-}
 
-export class Mining implements Mining {
+export class Mining {
   restURL: string
   constructor(restURL: string) {
     this.restURL = restURL
